utils/ai_helpers.py
"""
AI integration utilities for BUILDBTECH
"""
import os
import requests
from dotenv import load_load_dotenv

load_dotenv()

def generate_lab_report(experiment_title, observations, course):
    """Generate AI-powered lab report"""
    try:
        # Implement offline fallback if API is not available
        return f"Lab Report: {experiment_title}\n\nObservations: {observations}\n\nAnalysis will be generated when online."
    except Exception as e:
        return f"Error generating report: {str(e)}"

def generate_study_plan(subject, duration, goals):
    """Generate personalized study plan"""
    try:
        # Implement offline study plan generation
        return f"Study Plan for {subject}\nDuration: {duration}\nGoals: {goals}"
    except Exception as e:
        return f"Error generating study plan: {str(e)}"
