pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "pyinstaller>=6.12.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
